Refactor: 회원 가입을 MemberController로 이동
- 원래는 세션이 필요없는 기능은 Auth, 필요한 기능 Member 도메인에서 관리하려함
- 로그인, 로그아웃을 제외한 Member 리소스를 사용하는 모든 기능은 Member에서 관리할 예정
- 관련 테스트 코드의 임포트 수정
