+++ src/main/java/com/uranus/taskmanager/api/controller/MemberController.java
@@ -1,8 +1,17 @@
 package com.uranus.taskmanager.api.controller;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.uranus.taskmanager.api.request.SignupRequest;
+import com.uranus.taskmanager.api.response.SignupResponse;
+import com.uranus.taskmanager.api.service.MemberService;
+
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -11,12 +20,19 @@ import lombok.RequiredArgsConstructor;
 public class MemberController {
 	/**
 	 * Todo
-	 * (세션 or Token이 필요)
-	 * 로그아웃 - 로그인한 사용자의 세션 무효화 및 쿠키 삭제
+	 * 회원 가입 - 새로운 멤버 등록
 	 * 회원 정보 조회/수정
 	 * 비밀번호 찾기 - 가입한 이메일 통한 비밀번호 찾기 (세션 불필요)
 	 * 비밀번호 변경
 	 * 회원 탈퇴
 	 * 침여하고 있는 워크스페이스 목록 조회
 	 */
+	private final MemberService memberService;
+
+	@PostMapping("/signup")
+	public ResponseEntity<SignupResponse> signup(@Valid @RequestBody SignupRequest signupRequest) {
+		SignupResponse signupResponse = memberService.signup(signupRequest);
+		return ResponseEntity.status(HttpStatus.OK).body(signupResponse);
+	}
+
 }
