+++ src/main/java/com/uranus/taskmanager/api/service/MemberService.java
@@ -1,0 +1,25 @@
+package com.uranus.taskmanager.api.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.uranus.taskmanager.api.domain.member.Member;
+import com.uranus.taskmanager.api.repository.MemberRepository;
+import com.uranus.taskmanager.api.request.SignupRequest;
+import com.uranus.taskmanager.api.response.SignupResponse;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+
+	private final MemberRepository memberRepository;
+
+	@Transactional
+	public SignupResponse signup(SignupRequest signupRequest) {
+		Member member = signupRequest.toEntity();
+		// Todo: password 암호화
+		return SignupResponse.fromEntity(memberRepository.save(member));
+	}
+}
